+++ apps/fluent-tester/CHANGELOG.md
@@ -1,8 +1,62 @@
 # Change Log - @fluentui-react-native/tester
 
-This log was last generated on Thu, 05 May 2022 19:08:12 GMT and should not be manually modified.
+This log was last generated on Fri, 13 May 2022 18:55:00 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.61.1
+
+Fri, 13 May 2022 18:55:00 GMT
+
+### Patches
+
+- Bump @fluentui/react-native to v0.29.7
+
+## 0.61.0
+
+Wed, 11 May 2022 22:26:25 GMT
+
+### Minor changes
+
+- Dismiss and openOnHover (ruaraki@microsoft.com)
+- Bump @fluentui-react-native/menu to v0.6.0
+
+## 0.60.1
+
+Wed, 11 May 2022 19:33:59 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/menu to v0.5.2
+
+## 0.60.0
+
+Tue, 10 May 2022 20:49:25 GMT
+
+### Minor changes
+
+- Add doNotTakeCapture example with a hover callout test page entry (patboyd@microsoft.com)
+- Bump @fluentui-react-native/experimental-menu-button to v0.3.39
+- Bump @fluentui-react-native/menu to v0.5.1
+- Bump @fluentui/react-native to v0.29.6
+
+## 0.59.40
+
+Tue, 10 May 2022 18:33:23 GMT
+
+### Patches
+
+- Added font and color tokens (v.kozova13@gmail.com)
+- Bump @fluentui-react-native/experimental-avatar to v0.14.26
+
+## 0.59.39
+
+Tue, 10 May 2022 17:20:22 GMT
+
+### Patches
+
+- Changed badge appearance according to the spec, changed sizes to align with Web (v.kozova13@gmail.com)
+- Bump @fluentui-react-native/experimental-avatar to v0.14.25
 
 ## 0.59.38