+++ apps/win32/CHANGELOG.md
@@ -1,8 +1,56 @@
 # Change Log - @fluentui-react-native/tester-win32
 
-This log was last generated on Thu, 05 May 2022 19:08:12 GMT and should not be manually modified.
+This log was last generated on Fri, 13 May 2022 18:55:00 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.20.44
+
+Fri, 13 May 2022 18:55:00 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.61.1
+
+## 0.20.43
+
+Wed, 11 May 2022 22:26:25 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.61.0
+
+## 0.20.42
+
+Wed, 11 May 2022 19:33:59 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.60.1
+
+## 0.20.41
+
+Tue, 10 May 2022 20:49:25 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.60.0
+
+## 0.20.40
+
+Tue, 10 May 2022 18:33:23 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.59.40
+
+## 0.20.39
+
+Tue, 10 May 2022 17:20:22 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.59.39
 
 ## 0.20.38